refactor: move ExecAutoAnalyze into the util package
