+++ pkg/statistics/handle/autoanalyze/BUILD.bazel
@@ -8,22 +8,19 @@ go_library(
     deps = [
         "//pkg/domain/infosync",
         "//pkg/infoschema",
-        "//pkg/metrics",
-        "//pkg/parser/ast",
         "//pkg/parser/model",
         "//pkg/sessionctx",
         "//pkg/sessionctx/variable",
         "//pkg/statistics",
+        "//pkg/statistics/handle/autoanalyze/util",
         "//pkg/statistics/handle/lockstats",
         "//pkg/statistics/handle/logutil",
         "//pkg/statistics/handle/types",
         "//pkg/statistics/handle/util",
         "//pkg/types",
         "//pkg/util",
-        "//pkg/util/chunk",
         "//pkg/util/logutil",
         "//pkg/util/sqlescape",
-        "//pkg/util/sqlexec",
         "//pkg/util/timeutil",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
