+++ pkg/statistics/handle/autoanalyze/util/util.go
@@ -1,0 +1,73 @@
+package util
+
+import (
+	"time"
+
+	"github.com/pingcap/tidb/pkg/metrics"
+	"github.com/pingcap/tidb/pkg/parser/ast"
+	"github.com/pingcap/tidb/pkg/sessionctx"
+	"github.com/pingcap/tidb/pkg/statistics"
+	statslogutil "github.com/pingcap/tidb/pkg/statistics/handle/logutil"
+	statstypes "github.com/pingcap/tidb/pkg/statistics/handle/types"
+	statsutil "github.com/pingcap/tidb/pkg/statistics/handle/util"
+	"github.com/pingcap/tidb/pkg/util/chunk"
+	"github.com/pingcap/tidb/pkg/util/sqlescape"
+	"github.com/pingcap/tidb/pkg/util/sqlexec"
+	"go.uber.org/zap"
+)
+
+var execOptionForAnalyze = map[int]sqlexec.OptionFuncAlias{
+	statistics.Version0: sqlexec.ExecOptionAnalyzeVer1,
+	statistics.Version1: sqlexec.ExecOptionAnalyzeVer1,
+	statistics.Version2: sqlexec.ExecOptionAnalyzeVer2,
+}
+
+// ExecAutoAnalyze executes the auto analyze task.
+func ExecAutoAnalyze(
+	sctx sessionctx.Context,
+	statsHandle statstypes.StatsHandle,
+	sysProcTracker sessionctx.SysProcTracker,
+	statsVer int,
+	sql string,
+	params ...interface{},
+) {
+	startTime := time.Now()
+	_, _, err := execAnalyzeStmt(sctx, statsHandle, sysProcTracker, statsVer, sql, params...)
+	dur := time.Since(startTime)
+	metrics.AutoAnalyzeHistogram.Observe(dur.Seconds())
+	if err != nil {
+		escaped, err1 := sqlescape.EscapeSQL(sql, params...)
+		if err1 != nil {
+			escaped = ""
+		}
+		statslogutil.StatsLogger().Error(
+			"auto analyze failed",
+			zap.String("sql", escaped),
+			zap.Duration("cost_time", dur),
+			zap.Error(err),
+		)
+		metrics.AutoAnalyzeCounter.WithLabelValues("failed").Inc()
+	} else {
+		metrics.AutoAnalyzeCounter.WithLabelValues("succ").Inc()
+	}
+}
+
+func execAnalyzeStmt(
+	sctx sessionctx.Context,
+	statsHandle statstypes.StatsHandle,
+	sysProcTracker sessionctx.SysProcTracker,
+	statsVer int,
+	sql string,
+	params ...interface{},
+) ([]chunk.Row, []*ast.ResultField, error) {
+	pruneMode := sctx.GetSessionVars().PartitionPruneMode.Load()
+	analyzeSnapshot := sctx.GetSessionVars().EnableAnalyzeSnapshot
+	optFuncs := []sqlexec.OptionFuncAlias{
+		execOptionForAnalyze[statsVer],
+		sqlexec.GetAnalyzeSnapshotOption(analyzeSnapshot),
+		sqlexec.GetPartitionPruneModeOption(pruneMode),
+		sqlexec.ExecOptionUseCurSession,
+		sqlexec.ExecOptionWithSysProcTrack(statsHandle.AutoAnalyzeProcID(), sysProcTracker.Track, sysProcTracker.UnTrack),
+	}
+	return statsutil.ExecWithOpts(sctx, optFuncs, sql, params...)
+}
